% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anyDup.R
\name{anyDup}
\alias{anyDup}
\title{Find all duplicates, not just the first}
\usage{
anyDup(x)
}
\arguments{
\item{x}{A vector}
}
\description{
Using \code{duplicated}() will find the 2nd instance of a duplicate but not
all instances of a duplicate. \code{anyDup} returns all duplicates.
}
