% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Candidates}
\alias{Candidates}
\title{King County contributions to candidates and political committees in 2018}
\format{A data.frame with 39192 rows and 37 variables:
\describe{
  \item{filer_name}{Name of the candidate}
  \item{office}{The office they're running for}
  \item{legislative_district}{The legislative district of that office}
  \item{party}{The political party with which they're affiliated}
  \item{amount}{The dollar amount of the contribution}
  \item{contributor_name}{The name of the person who made the donation}
  ...
}}
\source{
\url{http://www.pdc.wa.gov}
}
\usage{
Candidates
}
\description{
A dataset containing the amount of money donated to candidates for political
office in King County, WA in 2018. For a complete list of all variables,
please see the data source.
}
\keyword{datasets}
