% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centerBin.R
\name{centerBin}
\alias{centerBin}
\title{Cut a set of numeric data into bins based on the middle value of the bin}
\usage{
centerBin(x, breaks)
}
\arguments{
\item{x}{A numeric string}

\item{breaks}{the desired breaks in that string set by what the middle value
should be}
}
\value{
Returns a vector of numeric data
}
\description{
This function cuts a vector of numeric data such that the breaks are set
using the middle value of a given bin rather than setting upper or lower
boundaries for each bin. For example, say you've got a vector of blood draw
times that weren't drawn at *exactly* the nominal time but were close to it.
Set the bins here to the nominal times, and the data will be filed into the
slot that's closest to the nominal time at which that sample should have been
drawn.
}
