% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ConcTime}
\alias{ConcTime}
\title{Example concentration-time data}
\format{
A data.frame with 1800 rows and 13 variables: \describe{
  \item{Drug}{Drug administered}

  \item{DoseRoute}{Route of dose administation: IV or SC}

  \item{SubjectID}{Unique identifier for the subejct}

  \item{k_meanpop}{Mean estimate of the terminal elimination rate constant k
  for the population}

  \item{Vd_meanpop}{Mean estimate of the volume of distribution for the
  population}

  \item{k}{Estimate of the terminal elimination rate constant k for this
  subject}

  \item{Vd}{Estimate of the volume of distribution for this subject}

  \item{Sex}{Sex of the subject}

  \item{Age}{Subject's age}

  \item{Occupation}{Subject's occupation}

  \item{TimeHr}{Time (hr) since dose}

  \item{Dose}{Dose of the drug (mg)}

  \item{Concentration}{Plasma drug concentration (ng/mL)}

  }
}
\usage{
ConcTime
}
\description{
Made up concentration-time data for a crossover study design in which
subjects received drugs A and B both IV and subcutaneously. For convenience,
fitted parameters and some metadata are included in the same data.frame,
although I don't like how replicative this data.frame is.
}
\keyword{datasets}
