% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Cotinine}
\alias{Cotinine}
\title{Example data for converting data.frames from long to wide format}
\format{
A data.frame with 44 rows and 9 variables:
\describe{
  \item{SampleID}{Sample ID}
  \item{File}{MS data file}
  \item{DateTime}{Date and time of the acquisition}
  \item{Method}{MS method file used}
  \item{Position}{Well position}
  \item{Analyte}{Analyte measured}
  \item{RT}{Retention time (min)}
  \item{Area}{Peak area (counts)}
  \item{Height}{Peak height (counts)}
}
}
\usage{
Cotinine
}
\description{
Example mass spec data for a run of the cotinine assay
}
\keyword{datasets}
