% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutNumeric.R
\name{cutNumeric}
\alias{cutNumeric}
\title{Cut a set of numeric data into bins and then set the new value to be the left
or right side of the bin}
\usage{
cutNumeric(x, breaks)
}
\arguments{
\item{x}{A numeric string}

\item{breaks}{the desired breaks in that string; these are the same as for
the base function "cut"}
}
\description{
Cut a set of numeric data into bins and then set the new value to be the left
or right side of the bin
}
