% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateConv.R
\name{dateConv}
\alias{dateConv}
\title{Convert dates from numbers to Date}
\usage{
dateConv(x)
}
\arguments{
\item{x}{A numeric string}
}
\description{
Since data imported from Excel often changes date and time data to numeric,
\code{dateConv} converts dates from decimal numbers to Date.
}
