% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillRep.R
\name{fillRep}
\alias{fillRep}
\title{Make a unique data.frame by combining NA values}
\usage{
fillRep(DF, dlplyBy, concatFields = NULL, returnDifs = FALSE)
}
\arguments{
\item{DF}{a data.frame that you want to remove replicates from}

\item{dlplyBy}{a character vector of the columns in DF that you want to use
to break down DF; it's what you're supplying to \code{plyr::dlply} as the
\code{.variables} parameter.}

\item{concatFields}{a character vector of columns that are likely to contain
replicates that you want to concatenate. For example, say you've got two
rows in a merged data.frame where one row has the value "File1" listed for
the column "DataSource", and the other has the value "File2", but
everything else is the same. If you list "DataSource" for
\code{concatFields}, then the result will condense those two original rows
into one row where the column "DataSource" contains the concatenated names
of the original two files separated by a space, e.g., "File1 File2".}

\item{returnDifs}{If set to TRUE, returns a list with the same name as the
original data.frame and where the contents of the list are: \enumerate{
\item a data.frame called "differences" that shows differences between all
the subsets of data that \code{fillRep} was not able to reduce. It will
have 0 rows if there are no problematic data. \item a data.frame called
"new data.frame" that is the new, unique data.frame.} If set to FALSE, the
default value, the differences will be printed but the output object will
be only the new, unique data.frame.}
}
\value{
Returns a unique data.frame or, if \code{returnDifs} is TRUE, a list
  of two data.frames.
}
\description{
This function takes a data.frame, looks for rows that are replicates that
differ only because one row has missing values where the other has values,
fills in all missing values and returns the unique (no replicates) data.frame
whenever possible. When replicates cannot be removed because there are
differences, fillRep returns a data.frame of those differences. Whenever the
function encounters a set of rows that will \emph{not} reduce, it prints the
problematic fields as a data.frame. Optionally, you can keep this data.frame
of problems by setting \code{returnDifs} to TRUE.
}
\examples{

# Concatenating anything that's in the column "SourceFile" if that's the
# only thing that differs:
fillRep(MyDF, dlplyby = c("SubjectID", "StudyDay"),
        concatFields = "SourceFile", returnDifs = TRUE)

# Not concatenating any fields; any non-unique values will result in
# multiple rows in the output data.frame:
fillRep(MyDF, dlplyby = "SubjectID")
}
