% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findAllDup.R
\name{findAllDup}
\alias{findAllDup}
\title{Find locations of all duplicates, not just the first}
\usage{
findAllDup(x)
}
\arguments{
\item{x}{A vector}
}
\description{
Using the base-R function \code{duplicated} will find the 2nd, 3rd, etc.
instance of a duplicate but not \emph{all} instances of a duplicate.
\code{findAllDup} returns the locations of all duplicates.
}
\examples{
findAllDup(c("Lightning", "Theo", "Kiki", "Lightning"))

}
