% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatXL_head.R
\name{formatXL_head}
\alias{formatXL_head}
\title{Save an Excel file with standard formatting for the header row}
\usage{
formatXL_head(DF, file, sheet = NA)
}
\arguments{
\item{DF}{input data.frame}

\item{file}{file name (character)}

\item{sheet}{sheet name (character). Defaults to the name of the supplied
data.frame if no other name is supplied.}
}
\value{
This does not return any R objects; instead, it saves an Excel file.

This does not return any R objects; it saves an Excel file.
}
\description{
This function saves a data.frame to an Excel file with the header row in bold
face, centered both horizontally and vertically, and with wrapped text. It
also sets the column widths to a reasonable guess based on the data you're
writing. For more details, please see \code{\link{formatXL}}.
}
\examples{
data(iris)

formatXL_head(DF = iris, file = "test.xlsx", sheet = "iris1")


}
