% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_color_hue function.R
\name{gg_color_hue}
\alias{gg_color_hue}
\title{Select colors evenly spaced around the color wheel}
\usage{
gg_color_hue(n)
}
\arguments{
\item{n}{The number of colors desired}
}
\description{
\code{gg_color_hue} takes as input the number of colors desired and outputs
colors such as the defaults in ggplot2 graphs. NB: I did not write this
function, and I don't recall where I got it.
}
