% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm_95CI.R
\name{gm_95CI}
\alias{gm_95CI}
\title{Calculate the geometric 95 percent confidence interval}
\usage{
gm_95CI(x, na.rm = TRUE, zero.propagate = FALSE)
}
\arguments{
\item{x}{A vector of numbers}

\item{na.rm}{Should NA values be removed? (logical)}

\item{zero.propagate}{Should zeroes be propagated? (logical)}
}
\value{
Returns a list of numbers
}
\description{
\code{gm_95CI} takes as input a numeric vector and returns the geometric 95
percent confidence interval
}
\details{
Because this calculation log-transforms the input data: \enumerate{\item values of zero
must be either omitted (zero.propagate = FALSE) or, if included (zero.propagate =
TRUE), the result will be NA, and \item the function is undefined for negative
numbers.}
}
\examples{
gm_95CI(rnorm(10, 5, 1))
gm_95CI(c(4, 3, 7, 19, 4, 0), zero.propagate = TRUE)
gm_95CI(c(4, 3, 7, 19, 4, 0), zero.propagate = FALSE)

}
