% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometricMean.R
\name{gm_sd}
\alias{gm_sd}
\title{Calculate the geometric standard deviation}
\usage{
gm_sd(x, na.rm = TRUE, zero.propagate = FALSE)
}
\arguments{
\item{x}{A vector of numbers}

\item{na.rm}{Should NA values be removed? (logical)}

\item{zero.propagate}{Should zeroes be propagated? (logical)}
}
\description{
\code{gm_sd} takes as input a numeric vector and returns the geometric
standard deviation
}
\details{
Per Wikipedia's entry on the geometric standard deviation,
\href{https://en.wikipedia.org/wiki/Geometric_standard_deviation}{https://en.wikipedia.org/wiki/Geometric_standard_deviation},
the geometric standard deviation is dimensionless. See that entry for the
formula used.

Also, note that the way you determine the range of the data for the geometric
mean and one geometric standard deviation is NOT \eqn{geometric mean value
+/- the geometric standard deviation}! Instead, it is \eqn{geometric mean /
geometric standard deviation} to \eqn{geometric mean * geometric standard
deviation}!
}
\examples{

gm_sd(rnorm(10, 5, 1))
gm_sd(c(5, 3, 6, 10, 2, 0), zero.propagate = TRUE)

}
