% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Metformin}
\alias{Metformin}
\title{Example data for data manipulation, graphing}
\format{
A data.frame with 37 rows and 6 variables:
\describe{
  \item{SampleID}{Sample ID}
  \item{Matrix}{biological matrix sample was in}
  \item{Metformin.in.working.solution..ng.uL.}{Concentration of metformin in working solution (ng/uL)}
  \item{Volume.working.solution.added..uL.}{Volume of working solution added to sample (uL)}
  \item{metformin.peak.area}{Metformin peak area (counts)}
  \item{d6.metformin.peak.area}{d6-metformin peak area (counts)}
}
}
\usage{
Metformin
}
\description{
Example mass spec data for a run of the metformin assay
}
\keyword{datasets}
