% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Oxycodone}
\alias{Oxycodone}
\title{Example data for data manipulation, graphing}
\format{
A data.frame with 12 rows and 3 variables:
\describe{
  \item{Time}{Time sample drawn post-dose (min)}
  \item{Conc1}{Concentration for 1st subject (ng/mL)}
  \item{Conc2}{Concentration for 2nd subject (ng/mL)}
}
}
\usage{
Oxycodone
}
\description{
Example oxycodone concentration-time data
}
\keyword{datasets}
