% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Pets}
\alias{Pets}
\title{Example data for joining data.frames}
\format{A data.frame with 29 rows and 2 variables:
\describe{
  \item{Name}{Student's name}
  \item{Number.of.pets}{Number of pets that student has}
}}
\usage{
Pets
}
\description{
Students' name and the number of pets they have for joining with 'Students'
data.frame.
}
\keyword{datasets}
