% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starSig function.R
\name{starSig}
\alias{starSig}
\title{Add stars to an appropriately rounded p value to indicate significance}
\usage{
starSig(pval, sig = 1, starlevels = c(0.001, 0.01, 0.05))
}
\arguments{
\item{pval}{A numeric vector of p values}

\item{sig}{The number of significant digits to include}

\item{starlevels}{A vector of the three values to use for significance marked
with ***, **, or *. Output: character vector of rounded numbers with
asterisks, e.g., c("0.03\*", "0.0002**", "0.9", "0.00008***")}
}
\description{
\code{starSig} checks whether the p value input is below a certain threshold,
rounds to the determined number of sig figs, and marks it with an asterisk or
asterisks as appropriate to its significance level. Output is a character
string.
}
\examples{
starSig(0.0235)
starSig(0.0235, 2)
starSig(0.5)
starSig(0.0004235)
}
