% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_comma.R
\name{str_comma}
\alias{str_comma}
\title{Collapse a vector of character strings into one readable string, separated by
commas and by "and"}
\usage{
str_comma(x, oxford = TRUE)
}
\arguments{
\item{oxford}{TRUE or FALSE for whether to use the Oxford comma}
}
\value{
Returns a single, collapsed character string
}
\description{
\code{str_comma} combines multiple character strings into one and makes them
more readable by separating those strings with commas and "and" as
appropriate. The place I had in mind for using this is in RMarkdown documents
where you want to insert a comment in the text that calls on a character
vector. For example, say you want to note in a text chunk which subjects in a
study have some measurement that is two standard deviations above the mean.
Rather than using only, e.g., \code{stringr::str_c(StudyOutliers, collapse =
", ")}, you could have it include the "and" in the appropriate location and
make your final document more polished, readable, and professional.
}
\examples{
str_comma(LETTERS[1:2])
str_comma(LETTERS[1:4])
str_comma(LETTERS[1:4], oxford = FALSE)


}
