% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_split_alt.R
\name{str_split_alt}
\alias{str_split_alt}
\title{Split a string of text by a specified pattern and return a vector of the
split text}
\usage{
str_split_alt(string, pattern, n = NA, retainPattern = FALSE)
}
\arguments{
\item{string}{The string of text to be split into pieces}

\item{pattern}{The pattern to use for splitting the text. The pattern will
not be retained in the output unless \code{retainPattern = TRUE}.}

\item{n}{The index of the output vector to be returned. If left as \code{NA},
this will return all items in the vector. If set to a specific number n, it
will return only the nth item of that vector. If set to "last", it will
return only the last item of the vector.}

\item{retainPattern}{TRUE or FALSE for whether to retain the pattern that was
used to separate the input text in the output vector. Warning: This was set
up for a situation in which the user is only looking to match a single,
exact pattern and \emph{not} a pattern with regular expressions in it. The
function will still work if you use regex, but the output won't be what
you're likely expecting.}
}
\value{
Returns a vector
}
\description{
This function works similarly to \code{stringr::str_split} except that it
returns a vector rather than a list of vectors. Also, you can have it return
only one item in that vector instead of the entire vector, which works better
when you're trying to extract a specified piece of text and fill it into a
column of a data.frame or tibble.
}
\examples{

# Applied to a vector of length 1:
MyString <- c("First filename.xlsx Second filename.xlsx")
str_split_alt(MyString, ".xlsx")
str_split_alt(MyString, ".xlsx", 1)
str_split_alt(MyString, ".xlsx", "last")
str_split_alt(MyString, ".xlsx", retainPattern = TRUE)

# Applied to a longer vector, e.g., a column in a data.frame:
MyDF <- data.frame(ColA = paste("Item", 1:4),
                   ColB = c("This is a string to be split.",
                            "This one, too",
                            "Also split this",
                            "Last, split this one."))

MyDF$ColC <- str_split_alt(MyDF$ColB, " ", "last")
MyDF$ColD <- str_split_alt(MyDF$ColB, " ", 1)
MyDF$ColE <- str_split_alt(MyDF$ColB, " ", 1, retainPattern = TRUE)


}
