% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Students}
\alias{Students}
\title{Example data for data manipulation, data checking, etc.}
\format{
A data.frame with 37 rows and 6 variables:
\describe{
  \item{Name}{First name}
  \item{Department}{Department}
  \item{Gender}{Gender}
  \item{Cookie}{Favorite type of cookie}
  \item{VampTV.hr}{Amount of time spent watching trashy vampire TV last week (hr)}
  \item{Sleep.hr}{Amount of sleep last night (hr)}
}
}
\usage{
Students
}
\description{
Students in the UW Medicinal Chemistry or Pharmaceutics Departments and
some made-up data on their imagined habits.
}
\keyword{datasets}
