% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeConv.R
\name{timeConv}
\alias{timeConv}
\title{Convert times from numbers to POSIXct}
\usage{
timeConv(x, tz = "UTC")
}
\arguments{
\item{x}{A numeric string}
}
\description{
Since data imported from Excel often changes date and time data to numeric,
\code{timeConv} converts dates from decimal numbers to Date with a time zone
of "UTC" unless otherwise specified.
}
