% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tukeyStar.R
\name{tukeyStar}
\alias{tukeyStar}
\title{ggplot2 boxplots of ANOVA results}
\usage{
tukeyStar(data, groupColumn, valueColumn, barsize = 1.5, textsize = 8,
  includeN = FALSE, returnStats = FALSE)
}
\arguments{
\item{data}{The input data.frame on which to perform an ANOVA}

\item{groupColumn}{The name of the column that contains the groups to be
compared (character).}

\item{valueColumn}{The name of the column that contains the values of
interest (character).}

\item{barsize}{The size of the bars to display in the graph. Default is 1.5.}

\item{textsize}{The size of the asterisks or other text to display in the
graph. Default is 8.}

\item{includeN}{Include a label with the number of observations for each
group? (TRUE or FALSE)}

\item{returnStats}{Return the overall p value from the ANOVA and the results
from the Tukey post-hoc test? If TRUE, output will be a list of the plot,
the overall p value, and the data.frame of the Tukey post-hoc comparisons.}
}
\value{
If returnStats is FALSE, then the output will be the ggplot2 style
  boxplot. If returnStats is TRUE, then the output will be a list of the
  ggplot2 style boxplot, the overall p value, and the data.frame of Tukey
  post-hoc stats.
}
\description{
\code{tukeyStar} makes a boxplot of data to be analyzed with an ANOVA and
draws brackets between comparisons found to be significant by a Tukey
post-hoc test.
}
\examples{
DF <- data.frame(Hen = rep(c("A", "B", "C"), each = 5), ChickWeight = rep(c(rnorm(5, mean = 5, sd = 1), rnorm(5, mean = 6, sd = 1), rnorm(5, mean = 4, sd = 1))))
tukeyStar(data = DF, groupColumn = "Hen", valueColumn = "ChickWeight")
tukeyStar(data = DF, groupColumn = "Hen", valueColumn = "ChickWeight", includeN = TRUE)
tukeyStar(data = DF, groupColumn = "Hen", valueColumn = "ChickWeight", returnStats = TRUE)

}
