% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UWsalaries}
\alias{UWsalaries}
\title{Salaries of University of Washington employees from 2013-2017 (public record)}
\format{
A data.frame with ~124,000 rows and 7 variables:
\describe{
  \item{Name}{Employee name}
  \item{JobTitle}{Job title}
  \item{Sal2013}{Salary in 2013}
  \item{Sal2014}{Salary in 2014}
  \item{Sal2015}{Salary in 2015}
  \item{Sal2016}{Salary in 2016}
  \item{Sal2017}{Salary in 2017}
}
}
\usage{
UWsalaries
}
\description{
The pre-tax salaries of all UW employees from 2013 to 2017 in US dollars. (NB:
Not all employees worked a full year or worked a full year in a given position.)
Why am I including this in this package? Because it's a great data set and also because 1. you have a right to
know as a taxpayer and 2. if you're a UW employee (and most of the people who will use this package are), you should know
that this information is part of the public record. Same thing for your
UW email, fyi. If you're interested in a titillating tale, ask me about a
*former* Spokane County Public Works Department supervisor who had an affair on company
time and wrote emails to his paramour, a supervisee, using his Spokane County
email. And then had his emails requested as part of a public record search by
the local newspaper.
}
\keyword{datasets}
