% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WALeg}
\alias{WALeg}
\title{Contributions to candidates for the Washington State Legislature for the 2020
election}
\format{
A data.frame with 38 variables: \describe{ \item{filer_name}{Name of
  the candidate} \item{office}{The office they're running for}
  \item{legislative_district}{The legislative district of that office}
  \item{party}{The political party with which they're affiliated}
  \item{amount}{The dollar amount of the contribution}
  \item{contributor_name}{The name of the person who made the donation} ... }
}
\source{
\url{https://data.wa.gov/Politics/Contributions-to-Candidates-and-Political-Committe/kv7h-kjye/data}.
Accessed on Aug. 1, 2020.
}
\usage{
WALeg
}
\description{
A dataset containing the amounts of money donated to candidates for the
Washington State legislature from three counties in Western Washington and
three counties in Eastern Washington as of August 1, 2020. For a complete
list of all variables and their descriptions, please see
\url{https://www.pdc.wa.gov/browse/open-data/contributions-candidates-and-political-committees}.
}
\keyword{datasets}
