% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whichDif.R
\name{whichDif}
\alias{whichDif}
\title{Determine which columns differ in a two-row data.frame}
\usage{
whichDif(x, ignore.na = TRUE)
}
\arguments{
\item{ignore.na}{}
}
\value{

}
\description{
This function takes a two-row data.frame and determines which cells differ.
For example, say you've got a data.frame of all the subjects in your study
and one subject is listed twice, even after you type \code{unique(MyDF)}.
There are dozens of columns in \code{MyDF}, so finding the column{s} that
differs would be tedious. This function will do that for you.
}
\examples{

whichDif(MyDF[MyDF$SubjectID == "Subject623", ])


}
