% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutNumeric.R
\name{cutNumeric}
\alias{cutNumeric}
\title{Cut numeric data into bins with vector output}
\usage{
cutNumeric(x, breaks)
}
\arguments{
\item{x}{A numeric string}

\item{breaks}{the desired breaks in that string; these are the same as for
the base function "cut". The number must be <= the bin into which it is
sorted and the number must be >= the next-lower bin.}
}
\value{
Returns a vector of numeric data
}
\description{
This function cuts a vector of numbers just like the base R function
\code{\link[base]{cut}} except that the output will be a vector of numbers
rather than that weird factor that's the default. The number listed in the
output will be the upper boundary of the bin for the number that was used in
the input.
}
\examples{

cutNumeric(c(-5, 17, 32, 65, 118), breaks = c(c(0, 15, 30, 60, 120)))


}
