% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutNumeric.R
\name{cutNumeric}
\alias{cutNumeric}
\title{Cut numeric data into bins with vector output}
\usage{
cutNumeric(x, breaks)
}
\arguments{
\item{x}{A numeric string}

\item{breaks}{the desired breaks in that string; these are the same as for
the base function "cut"}
}
\value{
Returns a vector of numeric data
}
\description{
This function cuts a vector of numbers just like the base R function "cut"
except that the output will be a vector of numbers rather than that weird
factor that's the default. The number listed in the output will be the lower
boundary of the bin for the number that was used in the input.
}
