% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateConv.R
\name{dateConv}
\alias{dateConv}
\title{Convert dates from numbers to Date}
\usage{
dateConv(x, dataSource = "Excel")
}
\arguments{
\item{x}{A numeric string}

\item{dataSource}{Either "Excel" (default), "Unix" or "R" depending on
where the data are coming from. Excel data have an origin of Dec. 30, 1899
whereas Unix and R data have an origin of Jan. 1, 1970.}
}
\value{
Returns a date object
}
\description{
Since data imported from Excel or simply manipulated within R often changes
date and time data to numeric, \code{dateConv} converts dates from decimal
numbers to Date.
}
