% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillRep.R
\name{fillRep}
\alias{fillRep}
\title{Make a unique data.frame by combining NA values}
\usage{
fillRep(DF, dlplyBy, concatFields = NULL, returnDifs = FALSE)
}
\arguments{
\item{DF}{a data.frame that you want to remove replicates from}

\item{dlplyBy}{a character string of the columns in DF that you want to use
to break down DF; it's what you're supplying to "plyr::dlply" as the
".variables" parameter.}

\item{concatFields}{a string of columns that are likely to contain replicates
that you want to concatenate. For example, say you've got two rows in a
merged data.frame where one row has the value "File1" listed for the column
"DataSource", and the other has the value "File2", but everything else is
the same. If you list "DataSource" for "concatFields", then the result will
be one row where the column "DataSource" contains the concatenated names of
the original two files.}

\item{returnDifs}{If set to TRUE, returns a list with the name of the
original data.frame where the contents of the list are 1. a data.frame
called "differences" that shows differences between all the subsets of data
that `fillRep` was not able to reduce. It will have 0 rows if there are no
problematic data. 2. a data.frame called "new data.frame" that is the new,
unique data.frame. If set to FALSE, the default value, the differences will
be printed but the output object will be only the new, unique data.frame.}
}
\description{
This function takes a data.frame, looks for rows that are replicates that
differ only because one row has missing values where the other has values,
fills in all missing values and returns the unique (no replicates) data.frame
whenever possible. When replicates cannot be removed because there are
differences, fillRep returns a data.frame of those differences. Whenever the
function encounters a set of rows that will *not* reduce, it prints the
problematic fields as a data.frame. Optionally, you can keep this data.frame
of problems by setting `returnDifs` to TRUE.
}
