% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findAllDup.R
\name{findAllDup}
\alias{findAllDup}
\title{Find locations of all duplicates, not just the first}
\usage{
findAllDup(x)
}
\arguments{
\item{x}{A vector}
}
\description{
Using \code{duplicated}() will find the 2nd instance of a duplicate but not
all instances of a duplicate. \code{anyDup} returns the locations of all
duplicates.
}
\examples{
findAllDup(c("Lightning", "Theo", "Kiki", "Lightning"))

}
