% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importChrom.R
\name{importChrom}
\alias{importChrom}
\title{Import an Agilent MassHunter chromatogram from a csv file into a useful
format}
\usage{
importChrom(csvfile, saveFile = FALSE)
}
\arguments{
\item{csvfile}{The csv file name that was exported from MassHunter Qual}
}
\value{
Output is a tidy data.frame with columns indicating the file, the
  chromatogram type, the ionization mode, the retention time, the counts,
  etc. If \code{saveFile} is set to TRUE, a new csv file of the tidied
  chromatographic data will be saved in the same directory, and it will have
  "- tidy" appended to the input file name.
}
\description{
This function takes a csv file that was generated by exporting TIC, SIM, MRM,
or binary-pump pressure chromatograms from Agilent MassHunter Qualitative
Analysis and tidies the data into a useful data.frame since the output from
Agilent software is so weird. Input is a character string that is the name of
the csv file.
}
