% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_sd.R
\name{mean_sd}
\alias{mean_sd}
\title{Calculate mean +/- sd or other parameters with appropriate sig figs}
\usage{
mean_sd(
  x,
  stdev.x = NULL,
  calcRange = FALSE,
  numDigRange = NA,
  calcCV = FALSE,
  calcMedian = FALSE,
  calc95CI = FALSE,
  reportn = FALSE,
  ndig = NA,
  na.rm = TRUE
)
}
\arguments{
\item{x}{A vector of numbers \emph{or} a single mean}

\item{stdev.x}{The sd when you want to provide it (leave as NULL if providing
a vector of numbers)}

\item{calcRange}{Should the range be reported? (logical) (TRUE or FALSE)}

\item{numDigRange}{Number of digits to report for the range}

\item{calcCV}{Should the coefficient of variation be reported? (logical)}

\item{calcMedian}{Should the median be reported? (logical)}

\item{calc95CI}{Should the 95\% confidence interval be reported? (logical) If
calc95CI is set to TRUE, it will return the 95\% confidence interval as
calculated as \emph{{\out{mean(x) +/- t<sub>n-1, 1-alpha/2</sub> *
sd(x)/sqrt(n)}}}}

\item{reportn}{Should the number of observations be reported? (logical)}

\item{ndig}{Optionally set the number of sig figs to use if you don't want
that calculated automatically}

\item{na.rm}{Should NA values be removed when calculating the mean, standard
deviation, etc.? (logical) They'll still be included in the count of
observations if \code{reportn} is TRUE.}
}
\value{
Returns a character string
}
\description{
\code{mean_sd} takes as input either a) a numeric vector or b) a number for
the mean and a second number for the sd, and it outputs the mean plus or
minus the standard deviation of that vector with an appropriate number of sig
figs. If a vector was supplied, optionally also obtain the median, range, CV,
and/or 95% confidence interval. Output is \emph{always} of class character.
}
\details{

}
\examples{
mean_sd(rnorm(10, 5, 1))
mean_sd(rnorm(10, 5, 1), calcRange = TRUE)
mean_sd(rnorm(10, 5, 1), calcCV = TRUE)
mean_sd(rnorm(10, 5, 1), calcMedian = TRUE)
mean_sd(rnorm(10, 5, 1), calc95CI = TRUE)
mean_sd(rnorm(10, 5, 1), reportn = TRUE)
mean_sd(rnorm(10, 5, 1), calcRange = TRUE, calcCV = TRUE, reportn = TRUE)

}
