% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_sd.R
\name{mean_sd}
\alias{mean_sd}
\title{Calculate mean +/- sd or other parameters with appropriate sig figs}
\usage{
mean_sd(x, stdev.x = NULL, calcRange = FALSE, numDigRange = NA,
  calcCV = FALSE, calcMedian = FALSE, calc95CI = FALSE,
  calcQuantiles = FALSE, reportn = FALSE, ndig = NA, na.rm = TRUE)
}
\arguments{
\item{x}{A vector of numbers OR a single mean}

\item{stdev.x}{The sd of the mean provided (leave as NULL if providing a
vector of numbers)}

\item{calcRange}{Should the range be reported? (logical) (TRUE or FALSE)}

\item{numDigRange}{Number of digits to report for the range}

\item{calcCV}{Should the coefficient of variation be reported? (logical)}

\item{calcMedian}{Should the median be reported? (logical)}

\item{calc95CI}{Should the 95\% confidence interval be reported? (logical) If
calc95CI is set to TRUE, it will return the 95\% confidence interval as
calculated as mean(x) +/- t~n-1, 1-alpha/2~ * sd(x)/sqrt(n)}

\item{calcQuantiles}{Should the quantiles be reported? (logical)}

\item{reportn}{Should the number of observations be reported? (logical)}

\item{ndig}{Optionally set the number of sig figs to use if you want to set
to a specific value}

\item{na.rm}{Should NA values be removed? (logical)}
}
\description{
\code{mean_sd} takes as input a numeric vector \emph{or} the mean and sd values and
outputs the mean plus or minus the standard deviation of that vector with an
appropriate number of sig figs. Output is ALWAYS of class character.
}
\details{
Example of ultimate output with all possible options set to TRUE:

5.1 (5) +/- 0.2 (4.7 to 5.3, 3.7\%, 95\%CI: 4.6 to 5.2, n = 18)

mean (median) +/- sd (range.min to range.max, CV\%, 95\%CI: lower to upper, n
= n).

NB: I have not set this up to deal with mean values in scientific notation.

NB: If sd(x) == 0, output will be unique(x).
}
\examples{
mean_sd(rnorm(10, 5, 1))
mean_sd(rnorm(10, 5, 1), calcRange = TRUE)
mean_sd(rnorm(10, 5, 1), calcCV = TRUE)
mean_sd(rnorm(10, 5, 1), calcMedian = TRUE)
mean_sd(rnorm(10, 5, 1), calc95CI = TRUE)
mean_sd(rnorm(10, 5, 1), reportn = TRUE)
mean_sd(rnorm(10, 5, 1), calcRange = TRUE, calcCV = TRUE, reportn = TRUE)

}
