% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noncompAUC.R
\name{noncompAUC}
\alias{noncompAUC}
\title{Calculate the AUC using the trapezoidal rule}
\usage{
noncompAUC(DF, concentration = "Concentration", time = "Time")
}
\arguments{
\item{DF}{Input data.frame with concentration-time data.}

\item{concentration}{The name of the column containing drug concentrations
(character).}

\item{time}{The name of the column containing time data (character).}
}
\value{
Returns a number
}
\description{
Given a data.frame of concentration-time data, \code{noncompAUC} calculates
the area-under-the-concentration-time curve for the times included in the
input data.frame using the linear trapezoidal rule. This function does
\emph{not} do any extrapolation to infinity.
}
