% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdCurve.R
\name{stdCurve}
\alias{stdCurve}
\title{Fit data to a standard curve}
\usage{
stdCurve(
  DF,
  rawPeak,
  rawIS,
  normPeak = NA,
  nominal,
  poly = "1st",
  weights = NULL,
  colorBy = NA
)
}
\arguments{
\item{DF}{The input data.frame with columns containing the nominal analyte
concentration and the instrument response}

\item{rawPeak}{The unadjusted instrument response column name (character).
Ignore this if data are already normalized by internal standard.}

\item{rawIS}{The internal standard column name (character). Ok to ignore if
data are already normalized by internal standard.}

\item{normPeak}{The name of the column containing instrument response
normalized by internal standard. Use this if the data are peak heights or
areas already divided by the IS peak height or area.}

\item{nominal}{The column with the nominal concentrations or masses
(character).}

\item{poly}{Should the data be fit to a 1st or 2nd order polynomial? Options:
"1st" or "2nd".}

\item{weights}{A vector of weights to use for the regression. If left as NA,
no weighting scheme will be used. Be careful that you don't have any
infinite values!}

\item{colorBy}{What column to color the points by in the standard curve
graph. If left as NA, points will all be black.}
}
\value{
Output is a list of the following named objects:

  1. Fit - A list of the fitted parameters

  2. CurvePlot - A plot of the data and the fitted line,

  3. Curve.DF - A data.frame of the points used for graphing the fitted line.

  4. Data - The original data with a column calculating the percent
  difference between the calculated and the nominal concentrations or masses
}
\description{
\code{stdCurve}, which was designed with mass spectrometry data in mind, fits
concentration and signal data to a standard curve and returns the calculated
betas, a plot of the standard curve, the data that were used to generate the
fitted line, and the original data that were used to generate the standard
curve.
}
