% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminalFit.R
\name{terminalFit}
\alias{terminalFit}
\title{Calculate the elimination rate of the terminal portion of a
concentration-time curve}
\usage{
terminalFit(
  DF,
  startValues = NA,
  concentration = "Concentration",
  time = "Time",
  tmax = NA,
  modelType = "monoexponential",
  returnDataUsed = FALSE,
  weights = NULL
)
}
\arguments{
\item{DF}{The data.frame with the concentration-time data}

\item{startValues}{list of starting values for A and k, NA if the start
values should be determined automatically. Coefficients for the
biexponential fit that will need values: A, alpha, B, and beta. For the
triexponential: A, alpha, B, beta, G, gamma. *An alternative approach:*
After having a lot of trouble with finding decent starting values for
triexponential decays using nls behind the scenes here, I've now changed
this function to optionally use the function nls2, which does a rigorous
search for starting values. For triexponential fits, if you submit a list
of starting values like usual, this function will use the regular nls
function. However, to use nls2 instead, set startValues to a two row
data.frame with columns for each coefficient in which the first row is the
minimum possible value to start using and the 2nd row is the maximum value
to start using for that coefficient. A warning: Because nls2 searches more
possible starting values, it is appreciably slower.}

\item{concentration}{A character string of the column name in DF that
contains concentration data}

\item{time}{A character string of the column name in DF that contains time
data}

\item{tmax}{The putative tmax, which is used for estimating the terminal
elimination rate. Time points \emph{before} tmax will be omitted from the
fit.}

\item{modelType}{The mathematical model to use for fitting the data; options
are "monoexponential", "biexponential", or "triexponential".}

\item{returnDataUsed}{Should the data used be returned? I wrote this script
initially for bootstrapping, where it can be useful to see what data were
used as input. For that reason, I'm including the option of returning the
data that were used.}

\item{weights}{Any weighting to be used for the nonlinear regression}
}
\description{
\code{terminalFit} fits concentration-time data to an exponential equation of
the form f(t) = A*exp(-k * t) where A is ~Cmax, k is the terminal elimination
rate constant, and t is time.
}
