% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminalFit.R
\name{terminalFit}
\alias{terminalFit}
\title{Calculate the elimination rate of the terminal portion of a
concentration-time curve}
\usage{
terminalFit(DF, startValues = NA, concentration = "Concentration",
  time = "Time", tmax = NA, returnDataUsed = FALSE, weights = NULL)
}
\arguments{
\item{DF}{The data.frame with the concentration-time data}

\item{startValues}{list of starting values for A and k, NA if the start
values should be determined automatically}

\item{concentration}{A character string of the column name in DF that
contains concentration data}

\item{time}{A character string of the column name in DF that contains time
data}

\item{tmax}{The putative tmax, which is used for estimating the terminal
elimination rate. Time points \emph{before} tmax will be omitted from the fit.}

\item{returnDataUsed}{Should the data used be returned? I wrote this script
initially for bootstrapping, where it can be useful to see what data were
used as input. For that reason, I'm including the option of returning the
data that were used.}

\item{weights}{Any weighting to be used for the nonlinear regression}
}
\description{
\code{terminalFit} fits concentration-time data to an exponential equation of
the form f(t) = A*exp(-k * t) where A is Cmax, k is the terminal elimination
rate constant, and t is time.
}
