% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminalFit.R
\name{terminalFit}
\alias{terminalFit}
\title{Calculate the elimination rate of the terminal portion of a
concentration-time curve}
\usage{
terminalFit(
  DF,
  startValues = NA,
  concentration = "Concentration",
  time = "Time",
  tmax = NA,
  modelType = "monoexponential",
  returnDataUsed = FALSE,
  weights = NULL,
  returnRSS = FALSE,
  useNLS_outnames = TRUE
)
}
\arguments{
\item{DF}{The data.frame with the concentration-time data}

\item{startValues}{list of starting values for A and k, NA if the start
  values should be determined automatically. \itemize{

  \item Coefficients for the monoexponential fit that will need values: A, k.

  \item Coefficients for the biexponential fit: A, alpha, B, and beta.

  \item For the triexponential: A, alpha, B, beta, G, gamma.

  \item \emph{An alternative approach:} After having a lot of trouble with
  finding decent starting values for triexponential decays using \code{nls}
  behind the scenes here, I've now changed this function to optionally use
  the function \code{nls2}, which does a more rigorous search for starting
  values than \code{nls}. (For more information, see the documentation on
  \code{\link[nls2]{nls2}} and the algorithm "\code{random-search}".) For
  triexponential fits, if you submit a list of starting values like usual,
  this function will use the regular \code{nls} function. However, to use
  \code{nls2} instead, set \code{startValues} to a two row data.frame with
  columns for each coefficient in which the first row is the minimum possible
  value to start using and the 2nd row is the maximum value to start using
  for that coefficient. Also, when \code{nls2} is used, the maximum number of
  iterations is set to increase from 50 (default) to 200. \strong{A warning:}
  Because nls2 searches more possible starting values, it can be appreciably
  slower.}}

\item{concentration}{A character string of the column name in DF that
contains concentration data}

\item{time}{A character string of the column name in DF that contains time
data}

\item{tmax}{The putative tmax, which is used for estimating the terminal
elimination rate. Time points \emph{before} tmax will be omitted from the
fit.}

\item{modelType}{The mathematical model to use for fitting the data; options
are "monoexponential", "biexponential", or "triexponential".}

\item{returnDataUsed}{Should the data used be returned? I wrote this script
initially for bootstrapping, where it can be useful to see what data were
used as input. For that reason, I'm including the option of returning the
data that were used.}

\item{weights}{Any weighting to be used for the nonlinear regression}

\item{returnRSS}{TRUE or FALSE for whether to resturn the residual sum of
squares. If set to TRUE, this will be the last column of the output
data.frame where all rows = the residual sum of squares. (I wanted the
output to still be a data.frame, so that's the place I could think of to
put it.)}

\item{useNLS_outnames}{TRUE or FALSE for whether to use the standard output
coeffecient names that come with the nls or nls2 functions, e.g.,
"Estimate", "Std. Error", "t value", and "Pr(>|t|)". These names are
annoying to work with for output data.frames b/c they don't follow standard
column-naming practices (they contain spaces and symbols). If set to FALSE,
the names of the output coefficient data.frame will be "Estimate", "SE",
"tvalue" and "pvalue".}
}
\value{
Returns a data.frame of the coefficients or a list containing a
  data.frame of the input data and a data.frame of the estimated coefficients
}
\description{
\code{terminalFit} fits an exponential equation to concentration-time data.
The equation should be of the form \eqn{f(t) = concentration = A * exp(-kt)}
where A is ~Cmax, k is the terminal elimination rate constant, and t is time,
and the model can be monoexponential, biexponential, or triexponential decay.
}
\examples{
# Example data to work with:
data(ConcTime)
Subj101 <- ConcTime \%>\% filter(SubjectID == 101 & DoseRoute == "IV" &
                                     Drug == "A") \%>\%
      select(SubjectID, TimeHr, Concentration)

# Automatically select the start values
terminalFit(Subj101, concentration = "Concentration", time = "TimeHr",
            modelType = "monoexponential")

# Set the start values yourself
terminalFit(Subj101, concentration = "Concentration", time = "TimeHr",
            tmax = 0, startValues = list(A = 30, k = 0.01),
            modelType = "monoexponential")

# Use the more robuse nls2 function to do the regression by selecting a range of
# values to search.
terminalFit(Subj101, concentration = "Concentration", time = "TimeHr",
            tmax = 0,
            startValues = data.frame(A = c(5, 50), k = c(0.0001, 0.05)),
            modelType = "monoexponential")

# Weight by 1/y
terminalFit(Subj101, concentration = "Concentration", time = "TimeHr",
            tmax = 0,
            weight = 1/Subj101$Concentration,
            modelType = "monoexponential")

# Get the residual sum of squares
terminalFit(Subj101, concentration = "Concentration", time = "TimeHr",
            modelType = "monoexponential", returnRSS = TRUE)

# Use better names for the columns in the output
terminalFit(Subj101, concentration = "Concentration", time = "TimeHr",
            modelType = "monoexponential", useNLS_outnames = FALSE)


}
