% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeConv.R
\name{timeConv}
\alias{timeConv}
\title{Convert times from numbers to POSIXct}
\usage{
timeConv(x, tz = "UTC", dataSource = "Excel")
}
\arguments{
\item{x}{A numeric string}

\item{dataSource}{dataSource Either "Excel" (default) or "Unix" depending on where
the data are coming from. Excel data have an origin of Dec. 30, 1899
whereas Unix or R data have an origin of Jan. 1, 1970.}
}
\value{
Returns a POSIXct object
}
\description{
Since data imported from Excel often changes date and time data to numeric,
\code{timeConv} converts dates from decimal numbers to Date with a time zone
of "UTC" unless otherwise specified.
}
\examples{
timeConv(43252.5)
# "2018-06-01 12:00:00 UTC"

timeConv(18526.2356, dataSource = "R")
# "2020-09-21 05:39:15 UTC"

}
