% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tukeyStar.R
\name{tukeyStar}
\alias{tukeyStar}
\title{ggplot2 boxplots of ANOVA results}
\usage{
tukeyStar(data, groupColumn, valueColumn, barsize = 1.5, textsize = 8,
  includeN = FALSE, returnStats = FALSE)
}
\arguments{
\item{data}{The input data.frame on which to perform an ANOVA}

\item{groupColumn}{The name of the column that contains the groups to be
compared (character).}

\item{valueColumn}{The name of the column that contains the values of
interest (character).}

\item{barsize}{The size of the bars to display in the graph. Default is 1.5.}

\item{textsize}{The size of the asterisks or other text to display in the
graph. Default is 8.}

\item{includeN}{Include the number of observations for each group? (TRUE or
FALSE)}

\item{returnStats}{Return the overall p value from the ANOVA and the results
from the Tukey post-hoc test? If TRUE, output will be a list of the plot,
the overall p value, and the data.frame of the Tukey post-hoc comparisons.}
}
\value{
If returnStats is FALSE, then the output will be the ggplot2 style
  boxplot. If returnStats is TRUE, then the output will be a list of the
  ggplot2 style boxplot, the overall p value, and the data.frame of Tukey
  post-hoc stats.
}
\description{
\code{tukeyStar} makes a boxplot of data that were compared with an ANOVA and
then evaluated with a Tukey post-hoc test.
}
\details{
INPUT: "data" is the data.frame you're plotting with ggplot2 "Tukey.df" is
the data.frame of results from as.data.frame(TukeyHSD(aov(Y ~ X))[[1]])
"valueColumn" is the column in "data" that the y axis is mapped to.
"groupColumn" is the corollary for the x axis. Title
}
