% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tukeyStar.R
\name{tukeyStar}
\alias{tukeyStar}
\title{ggplot2 boxplots of ANOVA results}
\usage{
tukeyStar(
  DF,
  groupColumn,
  valueColumn,
  barsize = 1.5,
  textsize = 8,
  includeN = FALSE,
  returnStats = FALSE
)
}
\arguments{
\item{DF}{The input data.frame on which to perform an ANOVA}

\item{groupColumn}{The column that contains the groups to be compared.}

\item{valueColumn}{The column that contains the values of interest.}

\item{barsize}{The size of the bars to display in the graph. Default is 1.5.}

\item{textsize}{The size of the asterisks or other text to display in the
graph. Default is 8.}

\item{includeN}{Include a label with the number of observations for each
group? (TRUE or FALSE)}

\item{returnStats}{Return the overall p value from the ANOVA and the results
from the Tukey post-hoc test? (TRUE or FALSE)}
}
\value{
If returnStats is FALSE, then the output will be the ggplot2 style
  boxplot. If returnStats is TRUE, then the output will be a list of:
  \describe{ \item{Overall_pvalue}{The p value for the overall ANOVA}
  \item{Comparisons}{A data.frame of the Tukey post-hoc comparisons}
  \item{Graph}{The boxplot graph}}
}
\description{
\code{tukeyStar} makes a boxplot of DF that could be analyzed by ANOVA and
draws brackets between comparisons found to be significant by a Tukey
post-hoc test.
}
\examples{
Chickens <- data.frame(Hen = rep(c("A", "B", "C"), each = 5),
                 ChickWeight = rep(c(rnorm(5, mean = 5, sd = 1),
                                     rnorm(5, mean = 6, sd = 1),
                                     rnorm(5, mean = 4, sd = 1))))

tukeyStar(DF = Chickens, groupColumn = Hen,
          valueColumn = ChickWeight)

tukeyStar(DF = Chickens, groupColumn = Hen,
          valueColumn = ChickWeight, includeN = TRUE)

tukeyStar(DF = Chickens, groupColumn = Hen,
          valueColumn = ChickWeight, returnStats = TRUE)

}
