% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whichDif.R
\name{whichDif}
\alias{whichDif}
\title{Determine which columns differ in a two-row data.frame}
\usage{
whichDif(x, ignore.na = TRUE)
}
\arguments{
\item{x}{A two-row data.frame}

\item{ignore.na}{TRUE or FALSE for whether to ignore NA values. Default is
TRUE. If \code{ignore.na = TRUE}, then one column with a missing value and
another column with a complete case will be counted as a match and the
missing value will be filled in with the complete one. If \code{ignore.na =
FALSE}, then one missing value and one complete case always counts as a
mismatch.}
}
\value{
a character vector of the column names where value differ
}
\description{
This function takes a two-row data.frame and determines which cells differ.
For example, say you've got a data.frame of all the subjects in your study
and one subject is listed twice, even after you type \code{unique(MyDF)}.
There are dozens of columns in \code{MyDF}, so finding the column{s} that
differs would be tedious. This function will do that for you.
}
\examples{
data(Pets)
whichDif(Pets[Pets$Name == "Eddie", ])


}
